requirement.txt
Django==4.1.3
celery==5.2.7
django-celery-beat==2.4.0
django-celery-results==2.4.0
djangorestframework==3.14.0
Pillow==9.3.0
redis==4.4.0
